# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MCG Building Violation Management System API
  description: REST API consumed by the MCG web portal module and the MCG HARYANA mobile app (building-violations module).
  version: 2.0.0
paths:
...
  /building-violations/api/cases/:
    get:
      tags:
      - building-violations
      summary: List Cases
      operationId: list_cases_building_violations_api_cases__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - building-violations
      summary: Create Case
      operationId: create_case_building_violations_api_cases__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseCreateIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /building-violations/api/cases/counts/:
    get:
      tags:
      - building-violations
      summary: Counts
      description: Badge counts for the inbox / home screen.
      operationId: counts_building_violations_api_cases_counts__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /building-violations/api/cases/{pk}/:
    get:
      tags:
      - building-violations
      summary: Case Detail
      operationId: case_detail_building_violations_api_cases__pk___get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - building-violations
      summary: Case Patch
      operationId: case_patch_building_violations_api_cases__pk___patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseDraftUpdateIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/submit/:
    post:
      tags:
      - building-violations
      summary: Submit
      operationId: submit_building_violations_api_cases__pk__submit__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemarksIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/ae_forward/:
    post:
      tags:
      - building-violations
      summary: Ae Forward
      operationId: ae_forward_building_violations_api_cases__pk__ae_forward__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AEForwardIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/ae_return/:
    post:
      tags:
      - building-violations
      summary: Ae Return
      operationId: ae_return_building_violations_api_cases__pk__ae_return__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemarksIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/issue_notice/:
    post:
      tags:
      - building-violations
      summary: Issue Notice
      description: JC issues any notice/order/memo (SCN, stop-work, sealing, demolition order, referral ...).
      operationId: issue_notice_building_violations_api_cases__pk__issue_notice__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueNoticeIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/record_service/:
    post:
      tags:
      - building-violations
      summary: Record Service
      operationId: record_service_building_violations_api_cases__pk__record_service__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordServiceIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/record_response/:
    post:
      tags:
      - building-violations
      summary: Record Response
      operationId: record_response_building_violations_api_cases__pk__record_response__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordResponseIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/ae_forward_response/:
    post:
      tags:
      - building-violations
      summary: Ae Forward Response
      operationId: ae_forward_response_building_violations_api_cases__pk__ae_forward_response__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AECommentIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/schedule_hearing/:
    post:
      tags:
      - building-violations
      summary: Schedule Hearing
      operationId: schedule_hearing_building_violations_api_cases__pk__schedule_hearing__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleHearingIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/record_hearing/:
    post:
      tags:
      - building-violations
      summary: Record Hearing
      operationId: record_hearing_building_violations_api_cases__pk__record_hearing__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordHearingIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/drop/:
    post:
      tags:
      - building-violations
      summary: Drop
      operationId: drop_building_violations_api_cases__pk__drop__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DropIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/record_appeal/:
    post:
      tags:
      - building-violations
      summary: Record Appeal
      operationId: record_appeal_building_violations_api_cases__pk__record_appeal__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordAppealIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/update_appeal/:
    post:
      tags:
      - building-violations
      summary: Decide Appeal
      description: 'Update the litigation flag: stay extended / vacated, hearing dates, later orders, final decision.'
      operationId: decide_appeal_building_violations_api_cases__pk__update_appeal__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppealIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/decide_appeal/:
    post:
      tags:
      - building-violations
      summary: Decide Appeal
      description: 'Update the litigation flag: stay extended / vacated, hearing dates, later orders, final decision.'
      operationId: decide_appeal_building_violations_api_cases__pk__decide_appeal__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppealIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/record_execution/:
    post:
      tags:
      - building-violations
      summary: Record Execution
      operationId: record_execution_building_violations_api_cases__pk__record_execution__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordExecutionIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /building-violations/api/cases/{pk}/close/:
    post:
      tags:
      - building-violations
      summary: Close
      operationId: close_building_violations_api_cases__pk__close__post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemarksIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /building-violations/api/cases/{pk}/timeline/:
    get:
      tags:
      - building-violations
      summary: Timeline
      operationId: timeline_building_violations_api_cases__pk__timeline__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: pk
        in: path
        required: true
        schema:
          type: string
          title: Pk
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AECommentIn:
      properties:
        response:
          type: integer
          title: Response
        comments:
          type: string
          title: Comments
      type: object
      required:
      - response
      - comments
      title: AECommentIn
    AEForwardIn:
      properties:
        remarks:
          type: string
          title: Remarks
          default: ''
        jc_user:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Jc User
        recommendation:
          type: string
          title: Recommendation
          default: ''
      type: object
      title: AEForwardIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RemarksIn:
      properties:
        remarks:
          type: string
          title: Remarks
          default: ''
      type: object
      title: RemarksIn
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /building-violations/api/auth/otp/verify/
servers:
- url: /
  description: Same origin (the module is mounted at /building-violations/)